Prefer react named exports (#33)
In this diff I change default and mixed react imports to named only.
From discussion with Dan Abramov I have a readmap for esm in react.
In react v17 there will be esm entry point with both named and default
export but docs and all components will be recommended to migrate
to named exports. I moving this initiative forward in a lot of modules.

Size also became slightly smaller because of simpler cjs interop.